New packet6.GetInnerMostPeerAddr to extract mac address from the innermost peer-address field of the relay message option.
Summary:
When you have multiple relay relaying messages you are going to end up with multiple relay message payloads stacked one on top of the other. this diff adds a new public function called GetInnerMostPeerAddr which provides the peer address fields of the relay information closest to the DHCP Client, that typically is set to the EUI-64 IPv6 address by the cloest relay to the client.
We currently use that to infer the mac address for a client, even if the client doesn't use DUID-LL[T].

Reviewed By: rb2k, malmond77, nemith

Differential Revision: D5592407

fbshipit-source-id: 92a1079cd3b49290c0921746b789f092b0934626