VendorData: Allow Juniper models with '-' in them
Summary:
Juniper, in their infinite wisdom, chose a vendor class string that
deliminates using the '-' character which is also found in thier model numbers.

This changes the logic to look for the first and last deliminator (hopefully
they never add dashes to serial numbers!) and everything in between is the
model number.

Reviewed By: pallotron

Differential Revision: D7873684

fbshipit-source-id: 5533a1221cd1dbc2d5654c8463cab08b2496df76