enhance juniper vendor option parsing
Summary:
Turns out that parsing Juniper VC get even more complicated.  There is
a potential for older version of Junos to not return a serial number inside the
VC and it is instead contained in the HostName option.

This takes a bit more catious appoach parsing the end of the string fist and if
no final seperator is found ('-') then it will place the hostname option as the
serial number.

Reviewed By: pallotron

Differential Revision: D7954668

fbshipit-source-id: 26f6cb8a202e7815a549f764c74d5968dec38784